+++ R/update.R
@@ -121,11 +121,11 @@ update_ <- function(object, data, refresh, ...) {
     call0 <- call
     call0$debug <- list(no_compile = TRUE, model_code = TRUE)
     recompile <- !identical(
-      eval(call0, env = e)$model_code, as.character(get_code(object))
+      eval(call0, envir = e)$model_code, as.character(get_code(object))
     )
   }
   if (!recompile) {
     call$debug <- list(stanfit = object$stanfit@stanmodel)
   }
-  eval(call, env = e)
+  eval(call, envir = e)
 }