+++ NEWS.md
@@ -5,6 +5,7 @@
   * Fixed an issue with the `update()` method in `lfo()`.
   * Fixed an issue with `"tau_alpha"` type parameters with the `as_draws()` method for categorical responses.
   * Fixed an issue with Stan code generation for models with time-varying covariates for categorical responses.
+  * Fixed an issue with `formula.dynamitefit()` when the model contained a `splines` component.
 
 # dynamite 1.3.1